# calc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Calculator Application",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5001",
      "description": "dev server (uses live data)"
    },
    {
      "url": "http://127.0.0.1:5001",
      "description": "prod server"
    }
  ],
  "paths": {
    "/add": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "addition",
        "description": "Will add two values",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "a": 1,
                  "b": 2
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/sub": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "substraction",
        "description": "Will substract two values",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "a": 1,
                  "b": 2
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/mul": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "multipication",
        "description": "Will multiply two values",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "a": 1,
                  "b": 2
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/div": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "division",
        "description": "Will division two values",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "a": 1,
                  "b": 2
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    }
  }
}